Additional level of abstraction for drawer activity
